build-and-deploy-docs/requirements.txt
mkdocs==1.2.3
jinja2==3.0.0
pyYAML==6.0.0
python-magic
